+++ docs/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - block only if standing or actively using the Block skill
 - separate resistance against acid
 - added Far Shot/Point Blank/Swipe/Charge skills
+- added Attract spell
 - skills/spells have experience counters
 - books give skill experience instead of a whole level
 - rebalanced spells (some spells removed, others damage modified, runes fire multiple times, etc...)